requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssmin==0.2.0
Flask==1.1.2
Flask-Assets==2.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
jsmin==2.2.2
lesscpy==0.14.0
MarkupSafe==1.1.1
numpy==1.19.0
pandas==1.0.5
ply==3.11
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
redis==3.5.3
requests==2.23.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.17
threadpoolctl==2.1.0
urllib3==1.25.9
webassets==2.0
Werkzeug==1.0.1
WTForms==2.3.1
